added keybind for pet food
