+++ .gitignore
@@ -1,4 +1,4 @@
 __pycache__/
 assets/models/
 .idea/
-.keybinds
+.settings/
