+++ gui_components/view/status.py
@@ -1,0 +1,29 @@
+import tkinter as tk
+from gui_components.interfaces import LabelFrame
+
+
+class Status(LabelFrame):
+    def __init__(self, parent, **kwargs):
+        super().__init__(parent, 'Status', **kwargs)
+
+        self.grid_columnconfigure(0, weight=1)
+        self.grid_columnconfigure(3, weight=1)
+
+        self.curr_cb = tk.StringVar()
+        self.curr_routine = tk.StringVar()
+
+        self.cb_label = tk.Label(self, text='Command Book:')
+        self.cb_label.grid(row=0, column=1, padx=5, pady=(5, 0), sticky=tk.E)
+        self.cb_entry = tk.Entry(self, textvariable=self.curr_cb, state=tk.DISABLED)
+        self.cb_entry.grid(row=0, column=2, padx=(0, 5), pady=(5, 0), sticky=tk.EW)
+
+        self.r_label = tk.Label(self, text='Routine:')
+        self.r_label.grid(row=1, column=1, padx=5, pady=(0, 5), sticky=tk.E)
+        self.r_entry = tk.Entry(self, textvariable=self.curr_routine, state=tk.DISABLED)
+        self.r_entry.grid(row=1, column=2, padx=(0, 5), pady=(0, 5), sticky=tk.EW)
+
+    def set_cb(self, string):
+        self.curr_cb.set(string)
+
+    def set_routine(self, string):
+        self.curr_routine.set(string)
