+++ src/modules/bot.py
@@ -1,6 +1,5 @@
 """An interpreter that reads and executes user-created routines."""
 
-
 import threading
 import time
 import cv2
